sb: source path and backing store debug enhancements
This commit enhances debugging with sb to provide better source path
substitutions to the debugger, as well as creating python modules for
gdb and lldb which are able to automatically load missing debug symbols
from the backing stores.

Previously, sb generated a very large list of subsititutions which
mapped a hard-coded cluster's build
area (/mathworks/devel/bat/CLUSTER/build) to the current sandbox.  It
did this for every cluster, and also for 5 additional possible site
names: AH, BLR, AHBAT, AHBATTEST, and LSBAT.  This was an attempt to
cover every possible path that could be used as the build area.  This
resulted in > 3000 path substitutions in the gdbinit file.  In addition
to a large number of unnecessary entries, build paths now often look
like

 /mathworks/devel/bat/filer/batfs0555-0/Bdastudio.352188.Bdacore_integ.351905.pass/build

which is not a predictable path.  So we now use the 'dwarfdump' tool to
read the debug symbol files and extract a list of the actual build paths
that were found in the debug symbols.  This results in a much more
compact and accurate list.

This is done for both the current sandbox and all of its backing stores.
Mapping the source paths for the backing stores allows us to load
missing debug symbols directly from the backing store job archive, and
ensure that the source file loaded by the debugger will exist and will
be the correct version.

The '-update-gdb-source-path' option for sb has been renamed to
'-update-debug-source-path', as this switch can now generate an lldbinit
file as well as a gdbinit file.  An additional option
'-[no-]debug-backing-stores' has been added.  This enables or disables
the loading of backing store debug symbols.

The source path mapping for sbccache sandboxes now expires 6 months
after the last use.  This prevents the list of sbccache build locations
from growing indefinitely.  Although it is possible these cached objects
can still be in ccache, it seems to be pretty unlikely, and not worth
the baggage of keeping this history forever.

* sb:
  - Rename -update-gdb-source-path to -update-debug-source-path and add
    to obsolete switches
  - Add -debug-backing-stores.  This controls whether to enable the
    backing store debug symbol loader when debugging.  The default value
    of this switch is '1' if all backing store sandboxes exist,
    otherwise it is 0.  This can be disabled with
    '-no-debug-backing-stores'.
  - Fix handling of MW_CUSTOM_DEVELOPER_DIR so that the requested lldb
    is used.
* perlmod/sb_sync/sb_sync_gdb_setup.pm:
  - Support creating lldbinit file as well as gdbinit
  - Rename .source-path-gdbinit file to
    .sbtools/sb-debug/source-path.gdbinit
  - Create .sbtools/sb-debug/source-path.lldbinit file
  - Run sbdwarfdump to generate source path list (a.k.a. compilation
    dir) for current sandbox as well as any source paths found in the
    backing stores.
  - Cache sbdwarfdump output in
    .sbtools/sb-debug/bat-build-locations.ARCH.json
    - If this file is not writable, cache the sbdwarfdump output in
      TEMPDIR using a file name based on the md5sum of the path to the
      sandbox, the cluster name, and the job number.
  - Track sandbox build locations in
    .sbtools/sb-debug/sandbox-build-locations.json
  - Functions to run sbdwarfdump during sbsync
  - Functions to load backing store debug symbol loader for gdb and lldb
* sbsync:
  - Generate .sbtools/sb-debug/bat-build-locations.ARCH.json in the
    background
  - Generate .sbtools/sb-debug/source-path.gdbinit (or .lldbinit on mac)
* sbcopy:
  - Record sandbox location history
  - Regenerate source-path gdbinit/lldbinit
* sbclone:
  - Record sandbox location history
  - Regenerate source-path gdbinit/lldbinit
* sbmake:
  - Record sandbox location history
* sbdwarfdump:
  - New tool -- wrapper around 'dwarfdump'
  - Reads debug strings from debug files in a sandbox and (optionally)
    its backing stores.  Returns a json data structure list the build
    roots found in the debug symbols as well as the build roots found in
    the backing store debug symbols.
* help/helpsrc/sbdwarfdump_expansions.pm:
* help/helpsrc/sbdwarfdump.pod:
  - sbdwarfdump help
* apps/runperl/sb_progarches.pm:
  - Add sbdwarfdump (Linux/Mac only)
* apps/cgir_tools/.gdbinit:
* apps/vim/mwtools/vimfiles/plugin/mwgdb.vim:
  - Update references to .source-path-gdbinit to new location
    .sbtools/sb-debug/source-path.gdbinit
* apps/gdb/python/mathworks/gdb_load_backing_store_symbols.py:
  - New gdb module
  - Responsible for loading debug symbols from backing stores
  - Adds parameters 'matlab-root' and 'backing-store-directories'
  - Adds command 'register-debug-objfile-loader'
* apps/lldb/python/mathworks/lldb_load_backing_store_symbols.py:
  - New lldb module
  - Responsible for loading debug symbols from backing stores
* apps/lldb/python/mathworks/mw_config_lldb.py:
  - Switch breakpoint callback for breaksegv from a command callback to
    a python callback.  This seems to fix a bug which was causing the
    MATLAB startup output from being hidden.
  - Update breaksegv command to run 'process handle ...' command
    immediately if MATLAB has already been started, otherwise preserve
    existing behavior of registering NSApplicationMain callback.
* perlmod/sb_ccache.pm:
  - Expire ccache sandbox location history after 6 months.  This
    prevents the list of sbccache sandbox build locations from growing
    indefinitely.
* help/helpsrc/sbccache.pod:
  - Document the above change
* perlmod/sb_getsandbox.pm:
  - Add GetClusterAndJobNumForSandbox.  This reliably returns the
    cluster name and job number for a sandbox whether it is a job
    archive, an sbsync sandbox, or an sbs sandbox.
* sbsrc/templates/.gdbinit_template:
  - Source apps/gdb/python/mathworks/gdb_load_backing_store_symbols.py
* sbsrc/templates/.lldbinit_template:
  - Source apps/lldb/python/mathworks/lldb_load_backing_store_symbols.py
* perlmod/sbsyscheck/sb_syscheck_validate.pm:
  - Add .lldbinit check

CR: ciolfi
RR: g1356058