vimrc: hardcode sbtools release external-apps area
Previously, this vimrc was using its current location as the starting
path to find external-apps.

This breaks down if the repo is cloned outside of SBTOOLS/apps/vim (a
goal).

It seems safe to hardcode the /hub/share/external-apps as the location
to look for Python. If a new version of python2/3 needs to be qualified
for the use of this vimrc, it can be done by exporting a $PATH with a
local copy of external-apps before the hub copy.