refactor: migrate gdb, mwtools to sub packages
By making gdb and mwtools both subpackages of this package, it enables
interactively installing vim-tools via :packadd once Vim is running.

Previously, the setup.vim file was simply adding mwtools/gdb to the rtp,
leaving the Vim startup routine to then find these new plugin/ dirs and
source their contents.

This works for `packadd` in the vimrc, but to enable `:packadd`
post-startup, we need setup.vim to actually source those sub-tools
itself (and what better way than with packages!).

So now, setup.vim adds vim-tools itself to the packpath, and uses
packadd to add the subtools.

NOTE: to install vim-tools via vimrc it is *necessary* to use `packadd!`
(with !) so that the setup.vim file is only sourced once.