Fix how termdebug goes from stopped to running.
termdebug.vim has a little state machine based on receiving messages
from gdb. It used to consider that the program is "running" as soon as
it sees the message "*running" on the communication channel. However,
with MATLAB threads, some random threads can start running when stopped
at a breakpoint (especially when you invoke the command "ppx" etc.). To
avoid confusion, made termdebug only consider the program running if it
recives the message '*running,thread-id="all"'