Fix "selected thread is running" error and BP polish
I noticed today during debugging that I was getting a bunch of "Selected
thread is running" error messages. Turns out that this was because we
were issuing -stack-info-frame from inside HandleCursor even for
*stopped events which are very short-lived (for instance when
temporarily interrupting the program for setting a BP). To prevent this,
I did two things:
1. Directly handle the *stopped event message instead of issuing
   -stack-info-frame. This seems to also prevent some unecessary
   flashing
2. I also only update the stack if the *stopped event message has a
   valid filename. This also prevents us from getting too many random
   delays

I also noticed that it is no longer necessary to interrupt the program
to actually set a BP :) I took that out and some initial bashing seems
to not reveal any problems.

CR: ppatil