Fix bugs with opening current frame
We were failing to navigate to the current line if there are no files
currently open in the editor. There was also a more serious bug where we
were invoking the new-ui mi command at the wrong time if we ever got a
hit-enter prompt and the user has 'ignorecase' set.